fix the ut
